requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Werkzeug==3.0.6
Jinja2==3.1.4
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.7
blinker==1.9.0
python-dotenv==1.0.1
gunicorn==23.0.0
email-validator==2.2.0
bcrypt==4.2.1
Flask-WTF==1.2.2
WTForms==3.2.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Pillow==11.0.0
python-dateutil==2.9.0
pytz==2024.2
requests==2.32.3
xhtml2pdf
reportlab